move buildingHours into its own folder
